Update date formatting and adjust image styles
- Change date format from short to long month representation in formatGameDate function for better readability.
- Increase maximum width of team images from 90px to 130px and maximum height of logos from 6.5em to 9.5em to enhance visual presentation.

+++ src/localize.js
@@ -57,19 +57,19 @@ export function formatGameDate(date, hass) {
     if (isToday) {
       return {
         weekday: localize("today", hass),
-        date: d.toLocaleDateString(language, { month: "short", day: "numeric" }),
+        date: d.toLocaleDateString(language, { month: "long", day: "numeric" }),
         time: d.toLocaleTimeString(language, { hour: "2-digit", minute: "2-digit" })
       };
     } else if (isTomorrow) {
       return {
         weekday: localize("tomorrow", hass),
-        date: d.toLocaleDateString(language, { month: "short", day: "numeric" }),
+        date: d.toLocaleDateString(language, { month: "long", day: "numeric" }),
         time: d.toLocaleTimeString(language, { hour: "2-digit", minute: "2-digit" })
       };
     } else {
       return {
         weekday: d.toLocaleDateString(language, { weekday: "long" }),
-        date: d.toLocaleDateString(language, { month: "short", day: "numeric" }),
+        date: d.toLocaleDateString(language, { month: "long", day: "numeric" }),
         time: d.toLocaleTimeString(language, { hour: "2-digit", minute: "2-digit" })
       };
     }
